Fix 32bit build
:Release Notes:
Fixed building for 32bit platform

:Detailed Notes:
Yet another forcing of type to linking with a correct pbnjson
function.
pbnjson::JValue have templated method declaration that accepts any
type, but specialized implementation exists only for a few of them.
It would be good to change that to templated generic implementation
(even if there will be kinda static_assert) in-place together with
forward declaration for specializations.

:Testing Performed:
Verified build for desktop and for bitbake (qemux86).

:QA Notes:

:Issues Addressed:
[GF-52639] possibility to replace NTP with other time source

Open-webOS-DCO-1.0-Signed-off-by: Nikolay Orliuk <nikolay.orliuk@lge.com>
Reviewed-on: https://g2g.palm.com/4160
Tested-by: Nikolay Orliuk <nikolay.orliuk@lge.com>

+++ Src/ClockHandler.cpp
@@ -272,7 +272,7 @@ bool ClockHandler::cbGetTime(LSHandle* lshandle, LSMessage *message, void *user_
 		{
 			reply = createJsonReply(true);
 			pbnjson::JValue offset = pbnjson::Object();
-			offset.put("value", it->second.systemOffset);
+			offset.put("value", (int64_t)it->second.systemOffset);
 			offset.put("source", system);
 			reply.put("offset", offset);
 			reply.put("utc", (int64_t)(time(0) + it->second.systemOffset));
